requirements.txt
Flask==2.2.3
openai==0.26.5
python-telegram-bot==13.7
better_profanity